requirements.txt
# Aggiungi qui le dipendenze, per esempio:
# requests==2.31.0
# numpy>=1.27
Flask>=2.0
python-dotenv>=1.0
